[C++] Split gandiva_jni into two dll, gandiva_jni and gandiva_protobuf
